Initial commit!
This has a fair amount of stuff in it. I start a server, listen on IPv4 and IPv6 (need to find out
how to use the same port on each...), and even have login options.

I need some post-login handling, command parsing, and a world to load, but this is a pretty sweet
start.